cleaned up and added comments. Moved structs to separate file for readability
