+++ src/client.go
@@ -6,119 +6,56 @@ import (
 	"math/rand"
 	"net/rpc"
 	"os"
-	"strconv"
 	"time"
 )
 
-type Player struct {
-	ID        int
-	Hand      []Card
-	Pairs     []Pairs
-	Opponents []Player
-}
-
-// ** adapted from mapreduce
-func GameServerSock() string {
-	s := "/var/tmp/824-gs-"
-	s += strconv.Itoa(os.Getuid())
-	return s
-}
-
-type Card struct {
-	Value string
-	Suit  string
-	used  bool
-}
-
-type Pairs struct {
-	One Card
-	Two Card
-}
-
-type JoinGameArgs struct {
-}
-
-type JoinGameReply struct {
-	Success bool
-	ID      int
-	// Hand  []Card
-	// Pairs []Pairs
-}
-
-type GameStatusRequest struct {
-}
-
-type GameStatusReply struct {
-	Complete      bool
-	Turn          int
-	CurrentPlayer int
-	Winner        int
-	Scores        []int
-	Players       []Player
-}
-
-type CardRequest struct {
-	Turn   int
-	Target int //Index of target player
-	Value  string
-}
-
-type CardRequestReply struct {
-	Turn   int
-	Cards  []Card
-	GoFish bool
-}
-
-type PlayPairRequest struct {
-	Turn  int
-	Owner int
-	Hand  []Card
-	Pair  []Pairs
-}
-
-type PlayPairReply struct {
-	Turn     int
-	Accepted bool
-}
-
-func Shuffle(slice []Card) {
-	r := rand.New(rand.NewSource(time.Now().Unix()))
-	for n := len(slice); n > 0; n-- {
-		randIdx := r.Intn(n)
-		slice[n-1], slice[randIdx] = slice[randIdx], slice[n-1]
-	}
-}
-
+// Main player functionality
 func (p *Player) PlayGoFish() {
+
 	var gameOver = false
 	for !gameOver {
 		var reply = callGetGameStatus()
+
 		if reply.Complete {
-			fmt.Printf("Game Done Client\n")
+			fmt.Printf("CLIENT: Game Over\n")
 			return
 		}
+
+		// update the player's game state information
 		p.Opponents = reply.Players
 		p.Hand = reply.Players[p.ID].Hand
 		p.Pairs = reply.Players[p.ID].Pairs
+
+		// Currently this player's turn
 		if reply.CurrentPlayer == p.ID {
 			p.doTurn()
 			p.endTurn()
 		}
+
 		gameOver = reply.Complete
+
 		time.Sleep(300 * time.Millisecond)
 	}
 }
 
+// Determine the card value and opponent to ask. Send RPC to the server
 func (p *Player) doTurn() {
+
 	r := rand.New(rand.NewSource(time.Now().Unix()))
+
+	// randomly choose an opponent to ask
 	var randIdx = p.ID
 	if len(p.Opponents) > 1 {
-		fmt.Printf("%v My hand %d\n", p.Hand, p.ID)
-		fmt.Printf("%v My paris %d\n\n\n\n", p.Pairs, p.ID)
+		fmt.Printf("Player %d my hand: %v\n", p.ID, p.Hand)
+		fmt.Printf("Player %d my pairs: %v\n\n", p.ID, p.Pairs)
+
 		for randIdx == p.ID {
 			randIdx = r.Intn(len(p.Opponents))
 		}
+
 		args := CardRequest{Target: p.Opponents[randIdx].ID}
+
+		// randomly choose a card value from your hand to ask for
 		if len(p.Hand) > 0 {
 			randIdx = r.Intn(len(p.Hand))
 			args.Value = p.Hand[randIdx].Value
@@ -127,22 +64,28 @@ func (p *Player) doTurn() {
 		}
 
 		p.callAskForCard(args)
+
 	} else {
 		time.Sleep(300 * time.Millisecond)
 	}
 }
 
+// Send RPC to the server asking for a card value from an opponent
 func (p *Player) callAskForCard(args CardRequest) {
 	reply := CardRequestReply{}
 	call("GameServer.AskForCards", &args, &reply)
+
+	// append any returned cards to your hand
 	if len(reply.Cards) > 1 || reply.Cards[0].Value != "-1" {
 		p.Hand = append(p.Hand, reply.Cards...)
 	}
 }
 
+// Find matching pairs in your hand, remove them and send to the server
 func (p *Player) endTurn() {
+
+	// find matching pairs
 	var pairList []Pairs
-	//Booo selection sort bad
 	for i := 0; i < len(p.Hand); i++ {
 		for x := i; x < len(p.Hand); x++ {
 			if p.Hand[i].Value == p.Hand[x].Value && !p.Hand[x].used && !p.Hand[i].used && i != x {
@@ -154,18 +97,22 @@ func (p *Player) endTurn() {
 
 	}
 
-	// Idk if this works YA YEET
+	// remove pairs from player's hand
 	var newList []Card
 	for _, v := range p.Hand {
 		if v.used != true {
 			newList = append(newList, v)
 		}
 	}
+
 	p.Hand = newList
+
+	// send RPC to server notifying of removed pairs and to let the next player go
 	p.callEndTurn(pairList)
 
 }
 
+// Send RPC to the server notifying of removed pairs and to let the next player go
 func (p *Player) callEndTurn(pairs []Pairs) {
 	args := PlayPairRequest{Owner: p.ID, Pair: pairs, Hand: p.Hand}
 	call("GameServer.EndTurn", &args, &PlayPairReply{})
@@ -173,13 +120,14 @@ func (p *Player) callEndTurn(pairs []Pairs) {
 	//Might want to update turn here from reply
 }
 
+// Send RPC to the server to get the current game state
 func callGetGameStatus() GameStatusReply {
 	reply := GameStatusReply{}
 	call("GameServer.AskGameStatus", GameStatusRequest{}, &reply)
 	return reply
 }
 
-// RPC - Join Game
+// Send RPC to the server asking to join game
 func callJoinGame() JoinGameReply {
 	args := JoinGameArgs{}
 	reply := JoinGameReply{}
@@ -189,13 +137,13 @@ func callJoinGame() JoinGameReply {
 	return reply
 }
 
+// Create the new player object
 func createPlayer() Player {
 	fmt.Println("CLIENT: successfully created player...")
 
 	// store player state
 	me := Player{}
 
-	// keep track of cards locally. if player goes down, server also keeps track
 
 	// join game via RPC
 	fmt.Println("CLIENT: joining game...")
@@ -205,11 +153,9 @@ func createPlayer() Player {
 		fmt.Println("Error: Could not join game.")
 		os.Exit(1)
 	}
-	// fmt.Println("Joined game")
 
 	me.ID = reply.ID
-	//me.Hand = reply.Hand
-	// me.Pairs = reply.Pairs
+
 	return me
 
 }
@@ -221,7 +167,6 @@ func runClient() {
 
 }
 
-// ** copied from mapreduce **
 // send an RPC request to the master, wait for the response.
 // usually returns true.
 // returns false if something goes wrong.
